+++ src/modules/api.ts
@@ -1,5 +1,4 @@
 import { AuthType } from '../types/api';
-import { components, operations } from '../types/public-api/schema';
 import { InternalAPI } from './internalApi';
 import { Consumers } from './consumers';
 import { Syncs, SyncsAPI } from './syncs';
@@ -35,29 +34,5 @@ export class API {
             throw new Error('Missing mandatory auth parameters');
         }
         this.internalApi = new InternalAPI(this.auth);
-    };
-
-    getMcpToken = async (
-        consumerId: components['schemas']['MCPAuthItem']['consumerId']
-    ): Promise<
-        operations['generate_mcp_token_mcp_token_post']['responses'][200]['content']['application/json']
-    > => {
-        if (!this.internalApi) {
-            await this._setup();
-        }
-        const body: components['schemas']['MCPAuthItem'] = {
-            clientId: this.auth.clientId,
-            clientSecret: this.auth.clientSecret,
-            accountId: this.auth.accountId,
-            consumerId,
-        };
-        if (this.auth.envId) {
-            body.envId = this.auth.envId;
-        }
-        if (this.auth.marketplaceId) {
-            body.marketplaceId = this.auth.marketplaceId;
-        }
-        const { data } = await this.internalApi.post('/mcp-token', body);
-        return data;
     };
 }